Add files via upload
This dataset collects information from 100k medical appointments in Brazil and is focused on the question of whether or not patients show up for their appointment. A number of characteristics about the patient are included in each row.

‘ScheduledDay’ tells us on what day the patient set up their appointment.
‘Neighborhood’ indicates the location of the hospital.
‘Scholarship’ indicates whether or not the patient is enrolled in the Brasilian welfare program Bolsa Família.
We need to investigate What factors are important for us to know in order to predict if a patient will show up for their scheduled appointment?